+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.3.3] - 2025-10-01
+
+### Fixed
+
+- Dialectic Streaming to follow SSE conventions
+- Sentry tracing in the deriver
+
 ## [2.3.2] - 2025-09-25
 
 ### Added